deploy/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
flask==3.0.3
gunicorn==23.0.0
